Set up cross-project for JVM implementation
This commit changes the directory structure and build settings to make
cross-building possible, but does not introduce Scala.js cross-building
yet.

Future commits will duplicate the JVM implementation to a Scala.js
specific directory. To ensure that the JVM sources are the ones that
git considers moved, and the JS sources are considered inserted, this
commit moves the JVM sources without inserting JS sources. This will
preserve the git blame for the JVM sources.