Fix NullPointerException tests on Scala.js
The semantics on Scala.js are different than on the JVM. For instance,
many cases involving null pointers are considered undefined behavior in
Scala.js. The JVM implementation relies on JVM-specific handling of
null pointers to throw NullPointerExceptions. In Scala.js, however, we
cannot rely on the platform to throw these, so they must be checked
manually.