Support a flag to enable always emitting the discriminator field (#1148)
Support a flag to enable always emitting the discriminator field

This commit adds the `alwaysEmitDiscriminator` flag to CodecMakerConfig
in order to support always emitting the discriminator field for ADT
leaf classes. This new flag has no effect on generated decoders, and it
does not effect encoder generation for classes that are not members of
ADTs. This new flag also should have no effect on codecs derived for the
root of the ADT.