Fix to remove needless allocations (performance) (#1335)
Before this fix, every decode operation would reallocate the `buf`
Array. This is because at the start of read, pos=head=0.

Then, because we've grown the buffer, at the end of the read operation
we'd allocate the array AGAIN to shrink it back because it is now larger
than the preferred buffer size config!

The problem started with the commit:

f0ebbe1e7  Code clean up (Andriy Plokhotnyuk) 2020-01-02|

```diff
-      newPos
-    } else {
-      if (tail > 0) buf = java.util.Arrays.copyOf(buf, buf.length << 1)
-      pos
-    }
+    } else buf = java.util.Arrays.copyOf(buf, buf.length << 1)
```

This change restores the original check for the `tail > 0` which is used
to indicate that the buffer is currently empty (start of the read
operation) and we should not grow it.